Migrating backend to Playbin2 and refactor GstPlayer
* CesarPlayer/CesarPlayer.mdp:
* CesarPlayer/Gui/PlayerBin.cs:
* LongoMatch/Gui/MainWindow.cs:
* CesarPlayer/Player/IPlayer.cs:
* CesarPlayer/gtk-gui/gui.stetic:
* CesarPlayer/Utils/MediaFile.cs:
* CesarPlayer/Player/GstError.cs:
* libcesarplayer/src/sources.gapi:
* CesarPlayer/Gui/VolumeWindow.cs:
* CesarPlayer/gtk-gui/objects.xml:
* CesarPlayer/Player/GstPlayer.cs:
* CesarPlayer/Handlers/Handlers.cs:
* CesarPlayer/Player/GstUseType.cs:
* CesarPlayer/MultimediaFactory.cs:
* CesarPlayer/Player/GstDVDEvent.cs:
* CesarPlayer/Utils/FramesCapturer.cs:
* CesarPlayer/Utils/IFramesCapturer.cs:
* CesarPlayer/Utils/IMetadataReader.cs:
* CesarPlayer/Player/GstAspectRatio.cs:
* CesarPlayer/Player/GstMetadataType.cs:
* CesarPlayer/Player/GstAudioOutType.cs:
* CesarPlayer/Player/GstVideoProperty.cs:
* libcesarplayer/src/bacon-video-widget.h:
* CesarPlayer/Handlers/StateChangedHandler.cs:
* CesarPlayer/gtk-gui/LongoMatch.Gui.PlayerBin.cs:
* libcesarplayer/src/bacon-video-widget-gst-0.10.c:
* CesarPlayer/gtk-gui/LongoMatch.Gui.VolumeWindow.cs:

+++ CesarPlayer/Handlers/Handlers.cs
@@ -26,7 +26,7 @@ namespace LongoMatch.Video.Handlers
 	public delegate void PlayListSegmentDoneHandler ();
 	public delegate void SegmentClosedHandler();
 	public delegate void SegmentDoneHandler();
-	public delegate void VolumeChangedHandler (int level);
+	public delegate void VolumeChangedHandler (double level);
 	public delegate void NextButtonClickedHandler ();
 	public delegate void PrevButtonClickedHandler ();
 	public delegate void ProgressHandler (float progress);
